Use SimulateConfig for CMRR
